add final enrichment
