+++ storehaus-algebra/src/main/scala/com/twitter/storehaus/algebra/MergeableEnirchment.scala
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2013 Twitter Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License. You may obtain
+ * a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.twitter.storehaus.algebra
+
+import com.twitter.algebird.{ Monoid, StatefulSummer }
+
+object MergeableEnrichment {
+  implicit def enrich[K, V](store: Mergeable[K, V]): EnrichedMergeable[K, V] =
+    new EnrichedMergeable[K, V](store)
+}
+
+class EnrichedMergeable[K, V](store: Mergeable[K, V]) {
+  def withSummer(summer: StatefulSummer[Map[K, V]]): Mergeable[K, V] =
+    Mergeable.withSummer(store, summer)
+
+  def unpivot[CombinedK, InnerK, InnerV](split: CombinedK => (K, InnerK))
+    (implicit ev: V <:< Map[InnerK, InnerV], monoid: Monoid[InnerV])
+      : Mergeable[CombinedK, InnerV] =
+    Mergeable.unpivot(store.asInstanceOf[MergeableStore[K, Map[InnerK, InnerV]]])(split)
+}
