+++ storehaus-algebra/src/test/scala/com/twitter/storehaus/algebra/BufferingStoreProperties.scala
@@ -23,7 +23,7 @@ import org.scalacheck.Properties
 object BufferingStoreProperties extends Properties("BufferingStore") {
   import StoreProperties.storeTest
   import MergeableStoreProperties.{ mergeableStoreTest, newStore }
-  import MergeableStoreAlgebra._
+  import MergeableStoreEnrichment._
 
   property("BufferingStore obeys the store properties") = storeTest {
     newStore[String, Map[Int, String]].withSummer(SummingQueue[Map[String, Map[Int, String]]](10))
