+++ storehaus-algebra/src/test/scala/com/twitter/storehaus/algebra/UnpivotedStoreProperties.scala
@@ -22,7 +22,7 @@ import org.scalacheck.Properties
 object UnpivotedStoreProperties extends Properties("UnpivotedStore") {
   import StoreProperties.storeTest
   import MergeableStoreProperties.{ newStore, mergeableStoreTest }
-  import MergeableStoreAlgebra._
+  import MergeableStoreEnrichment._
 
   property("UnpivotedStore obeys the store properties") = storeTest {
     val mergeableStore = MergeableStore.fromStore(new JMapStore[String, Map[Int, String]])
